seperate files accordingly
